Split quotes, typed quotes and quote processor
